code/EffableProject/requirements.txt
aiogram
python-dotenv
openai
SQLAlchemy>=2.0
asyncpg
alembic
greenlet
APScheduler